+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added an `OnDialogAdd` and `OnDialogSet` events to RumorState
 - Added the Clear command
-- Add an `OnChoose` event to Rumor
+- Add an `OnWaitForAdvance` and `OnWaitForChoose` event to Rumor
 
 ### Changed
 - AutoAdvance is now a time instead of a toggle, which indicates the amount of